requirements.txt
Flask
gunicorn
paho-mqtt
pymongo
python-dotenv